requirements.txt
# Core dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# ML and Video Processing
torch>=2.1.0
transformers>=4.37.0
qwen-vl-utils>=0.0.1
yt-dlp>=2024.1.0

# Database
supabase>=2.3.0
httpx>=0.26.0

# Optional but recommended
json-repair>=0.7.0

# JWT verification (Supabase tokens)
PyJWT>=2.8.0

# Development (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
ruff>=0.1.0
